requirements.txt
streamlit
pandas
fitz
pytesseract
Pillow